refactor: UserRepository에서 했던 중복 회원 검증을, UserService로 이동
