+++ src/main/java/com/kakao/cafe/repository/MemoryUserRepository.java
@@ -12,7 +12,6 @@ import com.kakao.cafe.domain.user.User;
 @Component
 public class MemoryUserRepository implements UserRepository {
 
-    private static final String DUPLICATED_USER_ID = "[ERROR] 존재하는 ID입니다. 다시 입력하세요.";
     private static final String NOT_FOUNDED_USER_ID = "[ERROR] 존재하지 않는 ID입니다.";
 
     private final List<User> users = Collections.synchronizedList(new ArrayList<>());
@@ -24,7 +23,6 @@ public class MemoryUserRepository implements UserRepository {
 
     @Override
     public void save(User user) {
-        duplicateUserIdCheck(user);
         users.add(user);
     }
 
@@ -38,18 +36,6 @@ public class MemoryUserRepository implements UserRepository {
     @Override
     public void deleteAll() {
         users.clear();
-    }
-
-    private void duplicateUserIdCheck(User inputUser) {
-        for (User user : users) {
-            checkIfTheNameIsSame(inputUser, user);
-        }
-    }
-
-    private void checkIfTheNameIsSame(User inputUser, User user) {
-        if (user.getUserId().equals(inputUser.getUserId())) {
-            throw new IllegalArgumentException(DUPLICATED_USER_ID);
-        }
     }
 
     public void update(String userId, User updateUser) {
