+++ src/test/java/com/kakao/cafe/repository/UserRepositoryTest.java
@@ -1,9 +1,7 @@
 package com.kakao.cafe.repository;
 
 import static org.assertj.core.api.Assertions.*;
-import static org.junit.jupiter.api.Assertions.*;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -55,20 +53,5 @@ class UserRepositoryTest {
 
         // then
         assertThat(user.get().getEmail()).isEqualTo("leejohy@naver.com");
-    }
-
-    @DisplayName("동일한 아이디로 가입하면 예외가 발생한다.")
-    @Test
-    void duplicate_id_error() {
-        User user1 = new User("lucid", "1234", "leejo", "leejohy@naver.com");
-        User user2 = new User("lucid", "0000", "leejo", "leej@naver.com");
-
-        userRepository.save(user1);
-
-        IllegalArgumentException exception = assertThrows(IllegalArgumentException.class,
-            () -> {
-                userRepository.save(user2);
-            });
-        assertThat(exception.getMessage()).isEqualTo("[ERROR] 존재하는 ID입니다. 다시 입력하세요.");
     }
 }
